make image builders easier to replace
