+++ plugins/docker_binary_builder/app/models/binary_builder.rb
@@ -141,7 +141,7 @@ class BinaryBuilder
   end
 
   def create_build_image
-    DockerBuilderService.build_docker_image(@dir, @output, dockerfile: DOCKER_BUILD_FILE, tag: image_name)
+    ImageBuilder.build_image(@dir, @output, dockerfile: DOCKER_BUILD_FILE, tag: image_name)
   end
 
   def docker_api_version
