+++ test/models/docker_builder_service_test.rb
@@ -396,65 +396,4 @@ describe DockerBuilderService do
       end
     end
   end
-
-  describe ".local_docker_login" do
-    run_inside_of_temp_directory
-
-    it "yields and returns" do
-      (DockerBuilderService.send(:local_docker_login) { 1 }).must_equal 1
-    end
-
-    describe "login commands" do
-      let(:called) do
-        all = []
-        DockerBuilderService.send(:local_docker_login) { |commands| all = commands }
-        all
-      end
-
-      before do
-        DockerRegistry.expects(:all).returns([DockerRegistry.new("http://fo+o:ba+r@ba+z.com")])
-        DockerBuilderService.class_variable_set(:@@docker_major_version, nil)
-      end
-
-      it "uses email flag when docker is old" do
-        DockerBuilderService.expects(:read_docker_version).returns("1.12.0")
-        called[1].must_equal "docker login --username fo\\+o --password ba\\+r --email no@example.com ba\\+z.com"
-      end
-
-      it "uses email flag when docker check fails" do
-        DockerBuilderService.expects(:read_docker_version).raises(Timeout::Error)
-        called[1].must_equal "docker login --username fo\\+o --password ba\\+r --email no@example.com ba\\+z.com"
-      end
-
-      it "does not use email flag on newer docker versions" do
-        DockerBuilderService.expects(:read_docker_version).returns("17.0.0")
-        called[1].must_equal "docker login --username fo\\+o --password ba\\+r ba\\+z.com"
-      end
-
-      it "can do a real docker check" do
-        called # checking that it does not blow up ... result varies depending on if docker is installed
-      end
-    end
-
-    it "copies previous config files from ENV location" do
-      File.write("config.json", "hello")
-      with_env DOCKER_CONFIG: '.' do
-        DockerBuilderService.send(:local_docker_login) do |commands|
-          dir = commands.first[/DOCKER_CONFIG=(.*)/, 1]
-          File.read("#{dir}/config.json").must_equal "hello"
-        end
-      end
-    end
-
-    it "copies previous config files from HOME location" do
-      Dir.mkdir(".docker")
-      File.write(".docker/config.json", "hello")
-      with_env HOME: Dir.pwd do
-        DockerBuilderService.send(:local_docker_login) do |commands|
-          dir = commands.first[/DOCKER_CONFIG=(.*)/, 1]
-          File.read("#{dir}/config.json").must_equal "hello"
-        end
-      end
-    end
-  end
 end
